fix(apps/frontend-manage): resolve rendering loop that blocked creation of case study elements
