+++ apps/frontend-manage/src/components/evaluation/hooks/useStackInstanceUpdates.tsx
@@ -20,6 +20,7 @@ function useStackInstanceUpdates({
         }
       })
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [activeInstance])
 }
