+++ apps/frontend-manage/src/components/questions/manipulation/options/SelectionCollectionOptions.tsx
@@ -10,12 +10,11 @@ import {
 import { useField } from 'formik'
 import { useTranslations } from 'next-intl'
 import Link from 'next/link'
-import { Dispatch, SetStateAction, useMemo } from 'react'
+import { Dispatch, SetStateAction, useEffect, useMemo } from 'react'
 import Select from 'react-select'
 import { twMerge } from 'tailwind-merge'
 import { ElementFormTypesSelection } from '../types'
 import AnswerCollectionInlineEditButton from './AnswerCollectionInlineEditButton'
-import useAnswerCollectionChangeEffect from './useAnswerCollectionChangeEffect'
 import useSelectAnswerCollectionOptions from './useSelectAnswerCollectionOptions'
 import useSelectedAnswerEntry from './useSelectedAnswerEntry'
 
@@ -67,11 +66,26 @@ function SelectionCollectionOptions({
   })
 
   // udpate the selected correct answers if the answer collection changes
-  useAnswerCollectionChangeEffect({
-    field: solutions,
-    helpers: solutionHelpers,
-    collectionAnswers,
-  })
+  useEffect(() => {
+    if (
+      !solutions.value ||
+      !collectionAnswers ||
+      collectionAnswers.length === 0
+    ) {
+      return
+    }
+
+    const newFieldValues = solutions.value.filter((id) =>
+      collectionAnswers.map((entry) => entry.value).includes(id)
+    )
+
+    solutionHelpers.setValue(newFieldValues)
+
+    // do not add value as a dependency --> rendering loop! - updates only on collection change desired
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [collectionAnswers])
+
+  console.log(solutions.value)
 
   // locally store the selected answer collection
   const selectedCollection = useMemo(() => {
