+++ apps/frontend-pwa/src/components/hooks/useRemainingInstances.ts
@@ -47,6 +47,7 @@ function useRemainingInstances({
       }
     }
     exec()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [quizId, instances, execution])
 }
